HARMONY-1416: Make sure /logs route forces EDL login, perform refactoring, and set up helpers for testing the logs route.
